fix(billing): satisfy Postgres CI and CodeQL
